tf2 modification and exp push

+++ 10-ColorBot/hvd-trans/utils.py
@@ -3,7 +3,9 @@ import  tensorflow as tf
 import  numpy as np
 from    tensorflow import keras
 import  urllib
+import horovod.tensorflow.keras as hvd
 
+from tensorflow.compat.v1 import string_split
 
 def parse(line):
     """
@@ -13,7 +15,7 @@ def parse(line):
     # Each line of the dataset is comma-separated and formatted as
     #    color_name, r, g, b
     # so `items` is a list [color_name, r, g, b].
-    items = tf.string_split([line], ",").values
+    items = string_split([line], ",").values
     rgb = tf.strings.to_number(items[1:], out_type=tf.float32) / 255.
     # Represent the color name as a one-hot encoded character sequence.
     color_name = items[0]
@@ -36,7 +38,8 @@ def maybe_download(filename, work_directory, source_url):
     if not tf.io.gfile.exists(work_directory):
         tf.io.gfile.makedirs(work_directory)
     filepath = os.path.join(work_directory, filename)
-    if not tf.io.gfile.exists(filepath):
+    # manual 
+    if hvd.rank() == 0 and not tf.io.gfile.exists(filepath):
         temp_file_name, _ = urllib.request.urlretrieve(source_url)
         tf.io.gfile.copy(temp_file_name, filepath)
         with tf.io.gfile.GFile(filepath) as f:
